backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
anytree==2.12.1
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.3.0
azure-core==1.31.0
azure-identity==1.19.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
boto3==1.9.147
botocore==1.12.253
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
cryptography==43.0.3
cymem==2.0.8
dataclasses-json==0.6.7
datasets==3.0.2
debugpy==1.8.1
decorator==5.1.1
deepeval==2.0.5
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
docutils==0.15.2
docx2txt==0.8
exceptiongroup==1.2.1
execnet==2.1.1
executing==2.0.1
fastapi==0.115.0
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.0.3
grobid-client==0.8.8
grobid-client-python==0.0.8
grobidmonkey==0.0.4
grpcio==1.63.2
h11==0.12.0
h5py==3.11.0
httpcore==0.15.0
httptools==0.6.1
httpx==0.23.0
huggingface-hub==0.26.1
idna==2.8
importlib-metadata==7.0.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.3.3
langchain==0.3.11
langchain-community==0.3.2
langchain-core==0.3.24
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langcodes==3.4.1
langsmith==0.1.134
language_data==1.2.0
latex2mathml==2.16.2
libclang==18.1.1
llama-cloud==0.1.2
llama-index==0.11.17
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.23
llama-index-embeddings-azure-openai==0.2.5
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-azure-openai==0.2.2
llama-index-llms-groq==0.3.0
llama-index-llms-openai==0.2.16
llama-index-llms-openai-like==0.3.0
llama-index-multi-modal-llms-openai==0.2.2
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.7
llamaindex-py-client==0.1.19
lxml==4.9.4
lz4==4.4.3
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
msal==1.31.0
msal-extensions==1.2.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.52.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==10.3.0
pinecone==5.4.0
pinecone-plugin-inference==2.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.3
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.5.2
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.1
pypdf==4.3.1
pysbd==0.3.4
pytest==8.3.3
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.18
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.2.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
s3transfer==0.2.1
safetensors==0.4.5
scipy==1.15.0
sentry-sdk==2.18.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.5
statsmodels==0.14.4
striprtf==0.0.26
tabulate==0.9.0
tenacity==8.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
thinc==8.3.2
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.1.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
websockets==13.1
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.16.0
xmltodict==0.14.1
xxhash==3.5.0
yarl==1.9.4
zipp==3.21.0
